:art: move menu to tools/
